requirements.txt
fastapi
uvicorn
python-multipart
email-validator
mongoengine
python-dotenv
pydantic-settings
python-jose[cryptography]
